Reduced memory footprint
Disabled 32 bit WAV support

+++ WaveReader/WaveReaderTest/SplitChannelsTest.cs
@@ -8,10 +8,10 @@ namespace WaveReaderTest
     public class SplitChannelsTest
     {
         private readonly int BitsPerByte = 8;
-        public int NumberOfChannels { get; set; }
-        public int BitsPerSample { get; set; }
+        public short NumberOfChannels { get; set; }
+        public short BitsPerSample { get; set; }
         public byte[] Input { get; set; }
-        public Int32[][] Output { get; set; }
+        public short[][] Output { get; set; }
         public int GetNumberOfFrames()
         {
             return (Input.Length * BitsPerByte) / (NumberOfChannels * BitsPerSample);
